Actually Working Build...

+++ user_admin.py
@@ -18,38 +18,53 @@ class Admin(Staff):
         self.connection.commit()
 
     def remove_staff(self, username):
-        query = f"DELETE FROM credentials WHERE username = '{username}'"
-        with self.cursor.execute(query):
-            print("Successfully deleted the user!")
-        self.connection.commit()
+        check = "SELECT * FROM credentials"
+        with self.cursor.execute(check):
+            row = self.cursor.fetchone()
+            while row:
+                if username in row:
+                    found = True
+                row = self.cursor.fetchone()
+
+        if found:
+            query = f"DELETE FROM credentials WHERE username = '{username}'"
+            with self.cursor.execute(query):
+                print("Successfully deleted the user!")
+            self.connection.commit()
+        else:
+            raise Exception()
 
     def change_permissions(self, username, new_perm):
-        query = f"UPDATE credentials SET permissions = '{new_perm}' WHERE username = '{username}'"
-        with self.cursor.execute(query):
-            print("Successfully altered user's permissions!")
-        self.connection.commit()
+        check = "SELECT * FROM credentials"
+        with self.cursor.execute(check):
+            row = self.cursor.fetchone()
+            while row:
+                if username in row:
+                    found = True
+                row = self.cursor.fetchone()
+
+        if found:
+            query = f"UPDATE credentials SET permissions = '{new_perm}' WHERE username = '{username}'"
+            with self.cursor.execute(query):
+                print("Successfully altered user's permissions!")
+            self.connection.commit()
+        else:
+            raise Exception()
 
     def view_user(self, username):
         data = None
         query = f"SELECT * FROM credentials WHERE username = '{username}'"
         with self.cursor.execute(query):
             data = self.cursor.fetchall()
-        return data if data != None else False
+        return data if len(data) > 0 else False
 
-    def add_flight(self, details):
-        journey_columns = [
-            "plane_id",
-            "departure_time",
-            "arrival_time",
-            "departing_from",
-            "arriving_to",
-        ]
+    def add_flight(self, details, columns):
         journey_details = []
         for k, v in details.items():
-            if k in journey_columns:
+            if k in columns:
                 journey_details.append(f"'{v}'")
 
-        flight_query = f"INSERT INTO plane ({', '.join(journey_columns)}) VALUES ({', '.join(journey_details)})"
+        flight_query = f"INSERT INTO journey_details ({', '.join(columns)}) VALUES ({', '.join(journey_details)})"
         with self.cursor.execute(flight_query):
             print("Succesfully added the plane!")
         self.connection.commit()
@@ -61,29 +76,29 @@ class Admin(Staff):
         self.connection.commit()
 
     def remove_flight(self, flight_ref):
-        query = f"DELETE FROM journey_details WHERE journey_id = '{flight_ref}'"
-        with self.cursor.execute(query):
-            print("Successfully deleted the flight!")
-        self.connection.commit()
+        check = "SELECT * FROM journey_details"
+        with self.cursor.execute(check):
+            row = self.cursor.fetchone()
+            while row:
+                if flight_ref in row:
+                    found = True
+                row = self.cursor.fetchone()
 
-    def add_plane(self, details):
-        plane_columns = [
-            "plane_type",
-            "plane_capacity",
-            "plane_size",
-            "fuel_capacity",
-            "speed",
-            "weight",
-            "seats_available",
-            "fuel_per_km",
-            "maintenance_date",
-        ]
+        if found:
+            query = f"DELETE FROM journey_details WHERE journey_id = '{flight_ref}'"
+            with self.cursor.execute(query):
+                print("Successfully deleted the flight!")
+            self.connection.commit()
+        else:
+            raise Exception()
+
+    def add_plane(self, details, columns):
         plane_details = []
         for k, v in details.items():
-            if k in plane_columns:
+            if k in columns:
                 plane_details.append(f"'{v}'")
 
-        plane_query = f"INSERT INTO plane ({', '.join(plane_columns)}) VALUES ({', '.join(plane_details)})"
+        plane_query = f"INSERT INTO planes ({', '.join(columns)}) VALUES ({', '.join(plane_details)})"
         with self.cursor.execute(plane_query):
             print("Succesfully added the plane!")
         self.connection.commit()
@@ -97,7 +112,18 @@ class Admin(Staff):
         self.connection.commit()
 
     def remove_plane(self, plane_ref):
-        query = f"DELETE FROM planes WHERE plane_id = '{plane_ref}'"
-        with self.cursor.execute(query):
-            print("Successfully deleted the plane!")
-        self.connection.commit()
+        check = "SELECT * FROM planes"
+        with self.cursor.execute(check):
+            row = self.cursor.fetchone()
+            while row:
+                if plane_ref in row:
+                    found = True
+                row = self.cursor.fetchone()
+
+        if found:
+            query = f"DELETE FROM planes WHERE plane_id = '{plane_ref}'"
+            with self.cursor.execute(query):
+                print("Successfully deleted the plane!")
+            self.connection.commit()
+        else:
+            raise Exception()
